+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### Unreleased
+### [v0.49](https://github.com/arez/arez-dom/tree/v0.49) (2020-01-16) · [Full Changelog](https://github.com/arez/arez-dom/compare/v0.48...v0.49)
 
 * Update the 'org.realityforge.arez' dependencies to version '0.166'